+++ indexers/blockindexer/kv/kv.go
@@ -12,10 +12,12 @@ import (
 
 	abci "github.com/tendermint/tendermint/abci/types"
 	"github.com/tendermint/tendermint/libs/pubsub/query"
-	"github.com/tendermint/tendermint/types"
 
-	"github.com/dymensionxyz/dymint/state/indexer"
+	indexer "github.com/dymensionxyz/dymint/indexers/blockindexer"
 	"github.com/dymensionxyz/dymint/store"
+	"github.com/dymensionxyz/dymint/types"
+
+	tmtypes "github.com/tendermint/tendermint/types"
 )
 
 var _ indexer.BlockIndexer = (*BlockerIndexer)(nil)
@@ -24,10 +26,10 @@ var _ indexer.BlockIndexer = (*BlockerIndexer)(nil)
 // events with an underlying KV store. Block events are indexed by their height,
 // such that matching search criteria returns the respective block height(s).
 type BlockerIndexer struct {
-	store store.KVStore
+	store types.KVStore
 }
 
-func New(store store.KVStore) *BlockerIndexer {
+func New(store types.KVStore) *BlockerIndexer {
 	return &BlockerIndexer{
 		store: store,
 	}
@@ -54,7 +56,7 @@ func (idx *BlockerIndexer) Has(height int64) (bool, error) {
 // primary key: encode(block.height | height) => encode(height)
 // BeginBlock events: encode(eventType.eventAttr|eventValue|height|begin_block) => encode(height)
 // EndBlock events: encode(eventType.eventAttr|eventValue|height|end_block) => encode(height)
-func (idx *BlockerIndexer) Index(bh types.EventDataNewBlockHeader) error {
+func (idx *BlockerIndexer) Index(bh tmtypes.EventDataNewBlockHeader) error {
 	batch := idx.store.NewBatch()
 	defer batch.Discard()
 
@@ -255,7 +257,7 @@ LOOP:
 			err        error
 		)
 
-		if qr.Key == types.BlockHeightKey {
+		if qr.Key == tmtypes.BlockHeightKey {
 			eventValue, err = parseValueFromPrimaryKey(it.Key())
 		} else {
 			eventValue, err = parseValueFromEventKey(it.Key())
@@ -479,7 +481,7 @@ func (idx *BlockerIndexer) match(
 	return filteredHeights, nil
 }
 
-func (idx *BlockerIndexer) indexEvents(batch store.Batch, events []abci.Event, typ string, height int64) error {
+func (idx *BlockerIndexer) indexEvents(batch types.StoreBatch, events []abci.Event, typ string, height int64) error {
 	heightBz := int64ToBytes(height)
 
 	for _, event := range events {
@@ -495,7 +497,7 @@ func (idx *BlockerIndexer) indexEvents(batch store.Batch, events []abci.Event, t
 
 			// index iff the event specified index:true and it's not a reserved event
 			compositeKey := fmt.Sprintf("%s.%s", event.Type, string(attr.Key))
-			if compositeKey == types.BlockHeightKey {
+			if compositeKey == tmtypes.BlockHeightKey {
 				return fmt.Errorf("event type and attribute key \"%s\" is reserved; please use a different key", compositeKey)
 			}
