+++ node/node.go
@@ -23,16 +23,16 @@ import (
 	"github.com/dymensionxyz/dymint/config"
 	"github.com/dymensionxyz/dymint/da"
 	daregistry "github.com/dymensionxyz/dymint/da/registry"
+	indexer "github.com/dymensionxyz/dymint/indexers/blockindexer"
+	blockidxkv "github.com/dymensionxyz/dymint/indexers/blockindexer/kv"
+	"github.com/dymensionxyz/dymint/indexers/txindex"
+	"github.com/dymensionxyz/dymint/indexers/txindex/kv"
 	"github.com/dymensionxyz/dymint/mempool"
 	mempoolv1 "github.com/dymensionxyz/dymint/mempool/v1"
 	nodemempool "github.com/dymensionxyz/dymint/node/mempool"
 	"github.com/dymensionxyz/dymint/p2p"
 	"github.com/dymensionxyz/dymint/settlement"
 	slregistry "github.com/dymensionxyz/dymint/settlement/registry"
-	"github.com/dymensionxyz/dymint/state/indexer"
-	blockidxkv "github.com/dymensionxyz/dymint/state/indexer/block/kv"
-	"github.com/dymensionxyz/dymint/state/txindex"
-	"github.com/dymensionxyz/dymint/state/txindex/kv"
 	"github.com/dymensionxyz/dymint/store"
 )
